Fix some warnings.

+++ src/gui.c
@@ -262,7 +262,7 @@ brgui_fm_result_t brgui_draw_file_manager(brgui_file_manager_t* state) {
           if (tab) {
             if (shift) --state->select_index;
             else       ++state->select_index;
-            if (state->select_index < 0 || state->select_index >= state->cur_dir.real_len) state->select_index = 0;
+            if (state->select_index < 0 || state->select_index >= (int)state->cur_dir.real_len) state->select_index = 0;
           }
         }
       brui_vsplit_end();
